added jeeves to route JS files
