+++ routes/api-routes.js
@@ -1,86 +1,6 @@
-const { v4: uuidv4 } = require("uuid");
 var db = require("../models");
 var passport = require("../config/passport");
-
-async function viewAllReservations(db) {
-  let results;
-  try {
-    results = await db.reservations.findAll({
-      include: [
-        {
-          model: db.users
-        },
-        {
-          model: db.users,
-          as: "instructor"
-        }
-      ]
-    });
-  } catch (err) {
-    console.log("Something went sideways", err);
-  }
-
-  let values = [];
-  results.forEach(res => {
-    values.push({
-      id: res.id,
-      start: res.start_Time,
-      end: res.end_Time,
-      instructor: res.instructor.first_name + " " + res.instructor.last_name,
-      customer: res.user.first_name + " " + res.user.last_name,
-      email: res.user.email,
-      status: res.status,
-      reservation_number: res.session_ID
-    });
-  });
-
-  return values;
-}
-async function viewAllUserReservations(db, req) {
-  let results;
-  try {
-    results = await db.reservations.findAll({
-      where: { userId: req.user.id },
-      include: [
-        {
-          model: db.users
-        },
-        {
-          model: db.users,
-          as: "instructor"
-        }
-      ]
-    });
-  } catch (err) {
-    console.log("Something went sideways", err);
-  }
-
-  let values = [];
-  results.forEach(res => {
-    values.push({
-      id: res.id,
-      start: res.start_Time,
-      end: res.end_Time,
-      instructor: res.instructor.first_name + " " + res.instructor.last_name,
-      customer: res.user.first_name + " " + res.user.last_name,
-      email: res.user.email,
-      status: res.status,
-      reservation_number: res.session_ID
-    });
-  });
-
-  return values;
-}
-async function createTimeSlot(db, req) {
-  let { Start_Time, End_Time, Id } = req;
-  let sesh_id = uuidv4();
-  db.reservations.create({
-    start_Time: Start_Time,
-    end_Time: End_Time,
-    instructorId: Id,
-    session_ID: sesh_id
-  });
-}
+var Jeeves = require("./reservation");
 
 module.exports = function(app) {
   // route for logging in a user
